let 2 players play "together" on the same testbed
This commit introduces a bit an inversion of control. Instead of Player 
objects doing the main iteration of play, this has now moved to the 
testbed. The testbed then calls play_one() on each of its players before 
doing the step update on its q values.

The graph for both players has been added in main()